Corrections and Comments
The "read mask image" function in the functional design document does not have to be available.
* This function was not included in the original specifications, but was added later as a request from Nara. Give priority to other function implementations.
.
It is a file selection dialog displayed by Brows, but please add "All (*. *)" To the filter.
Also, the filter Image files need only be * .jpg and * .png. (* .jpeg is also unnecessary)
* If you want to read * .bmp, select "All" and it will be displayed.
.
The window size cannot be changed.
If you want to click the exact position of the vertices, you will want to enlarge the image.
Is it difficult to make it resizable and magnify the edited image to fit the window?
.
If you want to respecify the image with "Browse" while editing the mask or after saving, reset the polygon editing state and the mask image.
⇒ Prepare a flag to determine whether it is the first reading or after the start of editing.
A confirmation dialog is displayed after the start of editing.
Example: "The mask being edited will be initialized. Are you sure?" (Yes, no)
.
If the extension at the time of saving is other than ".png", the Japanese of the error dialog should be as follows.
⇒ "Please specify png as the image format."


Please check below.


.
If you press "Save" again after "Save", it will end.
Large size of add / delete radio buttons